ironic: Restore check for unsupported soft power off
If the ironic driver does not support soft power off, we need to fall
back to hard power off. Due to an oversight in
67a27dc85996462b8f08348480072eec224e26e6 we stopped checking for this
and just returned a transient error in this case.

While returning a failure would have resulted in retrying with a hard
power off (albeit after reporting an error to the user), returning a
transient error means that it will retry the soft power off forever.

Restore the fallback and add tests to cover this case. Because we can't
set the dummy ironic server to return different responses on subsequent
calls we will always get an error, so check that it is from the fallback
hard power off.

(cherry picked from commit f75c6b7f0e0f4f65bacf5b98b2da7186001557c2)